new feature: computing filtering statistics
The number of pixels filtered due to each applied filter is printed to console. The filters are applied in order: if a pixel would be removed by a second filter, but it was already removed by a first,  the pixel is only counted in the pixels removed by the first filter. The implementation is temporary and will be substituted with some logging function in the future.