new feature: non-cumulative annotation dynamics
Annotation dynamics is now non-cumulative, meaning that the alpha distribution is divided into intervals and for each of them the log2 fold change of each annotation is computed with respect to the totality of the pixels. Still produces a dataframe which can be passed to the plot_annot_dynamics function.