bugfix: pandas.quantile crashing
Fixed pandas.quantile function crashing. The issue was due to passing a numpy.ndarray as an argument expected to be AnyArrayLike according to pandas; for pandas, an array like must be both iterable and implement __len__, the latter of which np.ndarray does not.  Solved by passing a list instead.