bugfix: fixed wrong fractions in HMM annotation
Fractions computed by the ann_fraction function did not use to sum to 1. This was due to how the portions of the intervals with no overlap where handled (basically discarded rather than assigned to the NaN annotation). Should be fixed now. Also added assertion that annotations for an interval sum up to 1 (though with some tolerance due to floating point issues).