new feature: introduced SparOperations
Sparsification is now treated just like normalization and filtering functions, that is, as an operation object to add to a processing flow. This is done to have a fully consistent way of passing processing parameters, as well as introducing new types of sparsification later on if need be. Also added p-value correction to sparsification scores (using Bonferroni) as an option.