enh: Operations now support a cleanup routine
Class OperationABC now supports the implementation of a cleanup routine. This is done through the cleanup method which is always executed after the run method. Does NOT clean up if the process generating the temporary file itself crashes, only if the used of said data does; that should be handled separately.