enh: improved OperationABC interface and logging
Made Operation interface more clear and tidy; the previous run method was moved to the process method, while the run method is now a wrapper for process and cleanup. This way the run method, which interfaces with the HiconaCooler object, is stable and can implement some checks on processing input and output. Moreover cleanup in called internally to the object.

Changed the specifics for the operations, that is, they must return all base columns of the table which they are provided, even if not used. This is to avoid a shift in column when saving to disk and resizing the table.

Improved progress logging by using actual logger with timestamps rather than print statements.