version: starting from scratch
Having understood what works and what does not, as well as what is needed, rewriting the entire package in a simpler, clearer way.

The old implementation was moved into legacy, where it will stay for backup purposes for now.  The old docs and test are moved too.

Meanwhile, in the main hicona folder, there is a new bare-bone implementation, allowing to create the basic objects, passing from one to the other, and extracting data from them.

The old features which are deemed necessary will be added back in progressively. Those with subpar results will be simply deprecated.  New tests and docs will be soon created.

Also, now added mypy for type checking and planning to add Github actions in the future.