refactor: PixelTable/BinTable replace HiconaTable
Deprecated HiconaTable in favor of PixelTable and BinTable classes. Advantages:

- returning and passing pandas and polars dataframes (full or in chunks) from and to many functions was becoming too messy (especially for typing and multiple flags). This way everything is more clean and dataframes and chunks can be obtained using the get_dataframe and get_chunks table method, respectively.
- a pixel table always stores a reference to the generating bin table. This allows for safe comparison across pixel tables (since bin ids depend on reference and chromosome order) as well as annotating pixels similarly to cooler annotate.
- table_ops and HiconaTable become obsolete and can be removed.

Also, instances of HiconaGraph do not store a copy of the bin table in memory, instead it is created from the graph on request. Although this is slower, it is more safe and it allows to retrieve eventual properties added on the graph. Tweaked the implementation of edge fetching to match this new implementation of node fetching.