fix: PixelTable.get_matrix indexing issue
When creating a matrix with PixelTable.get_matrix, the values used to fill the matrix might be floats. In that case, since an intermediate step is converting the polars df to numpy, the bin_ids were cast to float. This caused issues since then the indexes would be non-integers. Fixed by enforcing int conversion of indexes,
Also changed so that empty matrix values are NaNs rather than zeros.