feat: added HiconaGraph.hierarchical_clustering
Added a method to HiconaGraph to compute hierarchical clustering. This function adds to the graph posterior edge probability, reconstructed edges, edge clustering and node clustering as property maps. The interface is still unstable and will most likely be changed in the future to include entropy calculation, p and q estimates for the edges, and other information.

Also:
- self loops are now removed at graph creation since they interfere with the clustering algorithm
- changed some internal functions to work on iterables of edge/node descriptors rather than by assuming order which is not guaranteed
- other minor changes