enh: add_ind_col performs check of index presence
The function add_ind_col now checks whether the chunks in the iterable contain the index.
If all chunks already have it, the function just returns the chunks.
If no chunk has it, the index is added to all chunks (continuous across chunks)
If only some chunks have the index, raise an error.