rework: changed bin slicing and extent behavior
Having a partial genome binning in a BinTable, although more efficient for some operations, leads to a several issues, especially when  comparing PixelTables and plotting. For this reason, the BinTable class is now non-subsettable. Subsetting the PixelTable to a genomic region does not subset the associated BinTable.

Changed the BinTable.extent method to return the same interval as the one from Cooler.extent for a certain genomic region. Previously, the right boundary of the interval was included, now it is not (mimicking also how genomic coordinates work).

Added some quality of life attributes to the Table class.