rework: changed table tmp storage behavior
The Table class does not inherit from the temporary storage class any longer. Instead, an instance of the tmp storage is saved in an attribute, allowing to swap storage without making the Table obsolete. Also this prevents potentially dangerous overwrites on the same storage.
Fixed tests to account for annotation functions not returning new instances.