new: expected counts normalization
Added a new subpackage, misc, that will contain miscellaneous functions which might or might not be integrated in the main objects of the package later on.

Added three main functionalities: compute expected counts given a cooler file, use those as normalization on a pixel iterator, save a new cooler with normalized counts in the count column.