fix: palette deleting first pixel cluster
The palette used for plotting clustering in matrix mode used to assign the same color (white) to the background and the first cluster, therefore making it `vanish` from the matrix. This was due to a combination of plotting parameters and HicMatBase plot function rescaling values. Fixed by setting new proper parameters.

Also took the opportunity to distinguish between nan pixels (white) and 0 pixels (light gray).