fix: window mode stretching contact matrix
When plotting tracks in window mode, the retrieved data did not match the expected size. This was due to window mode trying to extend the region outside of chromosome boundaries, for which there is currently no check. Added pad of NaNs to the matrix to ensure the expected size. Maybe should be changed by bringing along chrom sizes.